requirements.txt
pandas==1.5.1
plotly==5.11.0
streamlit==1.14.0
Pillow==9.3.0

